refactor: Set up component structure according to the Atomic Design
