+++ .changeset/shiny-garlics-mate.md
@@ -1,0 +1,5 @@
+---
+"modern-walk": minor
+---
+
+refactor component structure
