+++ src/ui-core/components/atoms/CategoryTile/CategoryTile.types.tsx
@@ -1,0 +1,7 @@
+import { Category } from "../../../../enums/category";
+
+export interface CategoryTileProps {
+  category: Category.MensClothing | Category.WomensClothing;
+  url: string;
+  className: string;
+}
