+++ src/ui-core/components/atoms/CategoryTile/index.ts
@@ -1,0 +1,2 @@
+export { default as CateforyTile } from "./CategoryTile.component";
+export type { CategoryTileProps } from "./CategoryTile.types";
