+++ src/ui-core/components/atoms/Loading/Loading.types.tsx
@@ -1,0 +1,3 @@
+export interface LoadingProps {
+  message?: string;
+}
