+++ src/ui-core/components/atoms/Loading/index.ts
@@ -1,0 +1,2 @@
+export { default as Loading } from "./Loading.component";
+export type { LoadingProps } from "./Loading.types";
