+++ src/ui-core/components/atoms/Typography/Heading.types.tsx
@@ -1,0 +1,4 @@
+export interface HeadingProps {
+  variant: "h1" | "h2" | "h3" | "h4";
+  children?: React.ReactNode;
+}
