+++ src/ui-core/components/atoms/Typography/index.ts
@@ -1,0 +1,2 @@
+export { default as Heading } from "./Heading.component";
+export type { HeadingProps } from "./Heading.types";
