+++ src/ui-core/components/atoms/index.ts
@@ -1,0 +1,3 @@
+export * from "./CategoryTile";
+export * from "./Loading";
+export * from "./Typography";
