+++ src/ui-core/components/molecules/Navbar/Navbar.component.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "./navbar.css";
+import "./Navbar.component.css";
 
 export const Navbar = () => {
   return (
