+++ src/ui-core/components/molecules/Navbar/index.ts
@@ -1,0 +1,1 @@
+export { Navbar } from "./Navbar.component";
