+++ src/ui-core/components/molecules/ProductCard/ProductCard.component.css
