+++ src/ui-core/components/molecules/ProductCard/index.ts
@@ -1,0 +1,2 @@
+export { default as ProductCard } from "./ProductCard.component";
+export type { ProductCardProps } from "./ProductCard.types";
