+++ src/ui-core/components/molecules/index.ts
@@ -1,0 +1,2 @@
+export * from "./Navbar";
+export * from "./ProductCard";
