+++ src/ui-core/templates/index.ts
@@ -1,0 +1,2 @@
+export * from "./pages";
+export * from "./sections";
