+++ src/ui-core/templates/pages/index.ts
@@ -1,0 +1,4 @@
+export { MensClothing } from "./CategoryPage.template";
+export { WomensClothing } from "./CategoryPage.template";
+export { default as ProductList } from "./CategoryPage.template";
+export { default as Home } from "./Homepage.template";
