+++ src/ui-core/templates/pages/template.types.tsx
@@ -1,0 +1,11 @@
+import { Product } from "../../../models/Product.tsx";
+
+export interface ProductsProps {
+  url: string;
+  descriptionBackgroundColor: string;
+}
+
+export interface ProductListProps {
+  products: Product[];
+  descriptionBackgroundColor: string;
+}
