+++ src/ui-core/templates/sections/Categories/Categories.section.css
