+++ src/ui-core/templates/sections/Categories/index.ts
@@ -1,0 +1,1 @@
+export { default as Categories } from "./Categories.section";
