+++ src/ui-core/templates/sections/FlashSale/FlashSale.section.css
