+++ src/ui-core/templates/sections/FlashSale/index.ts
@@ -1,0 +1,1 @@
+export { default as FlashSale } from "./FlashSale.section";
