+++ src/ui-core/templates/sections/index.ts
@@ -1,0 +1,2 @@
+export * from "./Categories";
+export * from "./FlashSale";
